Upload first version of software
Two different software frameworks are added.

simGanESS: contains the simulation software based on nexus.
A proof-of-concept geometry is added with a single PMT (GaP1).

recoGanESS: Ganess REconstruction Software (gres, temporal name)
based on IC. It adds three different cities which imitate other
cities in IC but drop the usage of a tracking plane:
       - adelma transforms simulation output in waveforms (==buffy)
       - thekla simulates basic noise and creates pmaps (==hypathia)
       - theodora creates kdsts from pmaps (==dorothea)