GaP internal geometry PR3 changes.
Different naming for Solids, Physical Volumes...
Radius and length of PMT as variables.
Positions of PMT as vectors. X and Y coordinates separated. Position
is defined with the corresponfing X,Y pair.
Checked that Z position is correct.

Every volume with hole for the PMTs now grouped in a for loop (1 per
PMT).