Full geometry of GaP detector
The construction of the vessel is doen in a separated function called
"BuildVessel()". There the corresponding arms have been added to the
cylindric vessel. There are 4 arms, 2 for the HVFT and 2 for the
calibration sources and as vessel holders. This is LogicalVolume now.

The fucntion BuildTPC() will no longer have materials as input
arguments, materials are globally defined. The steel vessel arms have also
been filled with gas. High Voltage Feedthroughs has been added, one
that goes to the cathode and the other one to the gate.

In the vis.mac macros, corresponding colors have been defined.